Adding License
Choosing MIT-License, not really sure why but it feels to be best for my purposes.
I'm useless at programming, this is only a technical exercise so if you find this code, then you probably are looking in the wrong place. A general rule is to not use any of this code because it most likely will not serve your purpose, but feel very welcome to use at your leisure.